+++ packages/network-explorer/src/components/ClickableChecklistItem.tsx
