+++ packages/network-explorer/src/components/Table.tsx
@@ -11,8 +11,8 @@ import {
 	useReactTable,
 } from "@tanstack/react-table"
 import { useEffect, useState } from "react"
-import { fetchAndIpldParseJson } from "./utils.js"
-import useCursorStack from "./hooks/useCursorStack.js"
+import { fetchAndIpldParseJson } from "../utils.js"
+import useCursorStack from "../hooks/useCursorStack.js"
 import { WhereCondition } from "@canvas-js/modeldb"
 
 export type Column = {
