+++ packages/network-explorer/src/components/TableSelector.tsx
